catch non-json errors
There are instances where the error returned is not a json object, but a bunch of text. This is a hotfix to try to parse the response to text, then to json, and if it fails to parse to json, it would return the text instead.

There doesn't seem to be a way for us to check the type of response or to know anything about it until it is fully parsed. So I figured that parsing to text is safer than JSON and could always be converted to json.